Revert "Merge branch 'main' into Feature/dotnet6"
This reverts commit 08f5e8c540b651e46c151b87bb6e0def7c38f83c, reversing
changes made to 9188008f68b7cfe9f22205dd939caa481e73b0b1.

+++ fda-model-test/integrationtests/GraphicalUncertaintyShould.cs
@@ -54,9 +54,9 @@ namespace fda_model_test.integrationtests
             double binWidth = (int)(range / quantityOfBins);
             ThreadsafeInlineHistogram graphicalThreadsafeInlineHistogram = new ThreadsafeInlineHistogram(binWidth, convergenceCriteria);
             int masterseed = 1234;
-            Int64 progressChunks = 1;
-            Int64 _completedIterations = 0;
-            Int64 _ExpectedIterations = convergenceCriteria.MaxIterations;
+            int progressChunks = 1;
+            int _completedIterations = 0;
+            int _ExpectedIterations = convergenceCriteria.MaxIterations;
             if (_ExpectedIterations > 100)
             {
                 progressChunks = _ExpectedIterations / 100;
@@ -67,7 +67,7 @@ namespace fda_model_test.integrationtests
             {
                 seeds[i] = masterSeedList.Next();
             }
-            Int64 iterations = convergenceCriteria.MinIterations;
+            int iterations = convergenceCriteria.MinIterations;
             double lowerQuantile = 0.025;
             double upperQuantile = 0.975;
